+++ src/FreeFlight.CabinControl.App/Views/SettingsView.xaml.cs
@@ -1,6 +1,4 @@
 using System.Windows;
-using System.Windows.Controls;
-using FreeFlight.CabinControl.App.ViewModels;
 
 namespace FreeFlight.CabinControl.App.Views;
 
@@ -20,12 +18,4 @@ public partial class SettingsView
 
     private void CheckForUpdatesButton_Click(object sender, RoutedEventArgs e) =>
         CheckForUpdatesRequested?.Invoke(this, e);
-
-    private void FleetAccessKeyBox_PasswordChanged(object sender, RoutedEventArgs e)
-    {
-        if (DataContext is SettingsViewModel settings && sender is PasswordBox passwordBox)
-        {
-            settings.FleetApiAccessKey = passwordBox.Password;
-        }
-    }
 }